Fix and add integration tests for right & full joins
