+++ blackbox/docs/sql/joins.txt
@@ -83,32 +83,32 @@ Left Outer Joins
 like **Inner** join. Additionally it returns tuples for all other records from ​*left*​ that don't
 match any record on the ​*right*​ by using null values for the columns of the ​*right*​ relation::
 
-    cr> select e.name || ' ' || e.surname as employee, d.name as manager_of_department
+    cr> select e.name || ' ' || e.surname as employee, coalesce(d.name, '') as manager_of_department
     ... from employees e left join departments d
     ... on e.id = d.manager_id
     ... order by e.id;
-    +-------------------------------+-----------------------+
-    | employee                      | manager_of_department |
-    +-------------------------------+-----------------------+
-    | John Doe                      | Administration        |
-    | John Smith                    | IT                    |
-    | Sean Lee                      |                       |
-    | Rebecca Sean                  |                       |
-    | Tim Ducan                     |                       |
-    | Robert Duval                  |                       |
-    | Clint Johnson                 |                       |
-    | Sarrah Mcmillan               |                       |
-    | David Limb                    |                       |
-    | David Bowe                    |                       |
-    | Smith Clark                   | Marketing             |
-    | Ted Kennedy                   |                       |
-    | Ronald Reagan                 |                       |
-    | Franlkin Rossevelt            |                       |
-    | Sam Malone                    |                       |
-    | Marry Georgia                 |                       |
-    | Tim Doe                       | Human Resources       |
-    | Tim Malone                    | Purchasing            |
-    +-------------------------------+-----------------------+
+    +--------------------+-----------------------+
+    | employee           | manager_of_department |
+    +--------------------+-----------------------+
+    | John Doe           | Administration        |
+    | John Smith         | IT                    |
+    | Sean Lee           |                       |
+    | Rebecca Sean       |                       |
+    | Tim Ducan          |                       |
+    | Robert Duval       |                       |
+    | Clint Johnson      |                       |
+    | Sarrah Mcmillan    |                       |
+    | David Limb         |                       |
+    | David Bowe         |                       |
+    | Smith Clark        | Marketing             |
+    | Ted Kennedy        |                       |
+    | Ronald Reagan      |                       |
+    | Franklin Rossevelt |                       |
+    | Sam Malone         |                       |
+    | Marry Georgia      |                       |
+    | Tim Doe            | Human Resources       |
+    | Tim Malone         | Purchasing            |
+    +--------------------+-----------------------+
     SELECT 18 rows in set (... sec)
 
 Right Outer Joins
@@ -122,16 +122,16 @@ match any record on the ​*left*​ by using null values for the columns of the
     ... from employees e right join departments d
     ... on e.id = d.manager_id
     ... order by d.id;
-    +-------------------------------+-----------------------+
-    | employee                      | manager_of_department |
-    +-------------------------------+-----------------------+
-    | John Doe                      | Administration        |
-    | Smith Clark                   | Marketing             |
-    | Tim Malone                    | Purchasing            |
-    | Tim Doe                       | Human Resources       |
-    |                               | Shipping              |
-    | John Smith                    | IT                    |
-    +-------------------------------+-----------------------+
+    +-------------+-----------------------+
+    | employee    | manager_of_department |
+    +-------------+-----------------------+
+    | John Doe    | Administration        |
+    | Smith Clark | Marketing             |
+    | Tim Malone  | Purchasing            |
+    | Tim Doe     | Human Resources       |
+    |             | Shipping              |
+    | John Smith  | IT                    |
+    +-------------+-----------------------+
     SELECT 6 rows in set (... sec)
 
 Full Outer Joins
@@ -143,33 +143,33 @@ match any record on the ​*right*​ by using null values for the columns of th
 Additionally it returns tuples for all other records from ​*right*​ that don't match any record on
 the ​*left*​ by using null values for the columns of the ​*left*​ relation::
 
-    cr> select e.name || ' ' || e.surname as employee, d.name as manager_of_department
+    cr> select e.name || ' ' || e.surname as employee, coalesce(d.name, '') as manager_of_department
     ... from employees e full join departments d
     ... on e.id = d.manager_id
     ... order by e.id;
-    +-------------------------------+-----------------------+
-    | employee                      | manager_of_department |
-    +-------------------------------+-----------------------+
-    | John Doe                      | Administration        |
-    | John Smith                    | IT                    |
-    | Sean Lee                      |                       |
-    | Rebecca Sean                  |                       |
-    | Tim Ducan                     |                       |
-    | Robert Duval                  |                       |
-    | Clint Johnson                 |                       |
-    | Sarrah Mcmillan               |                       |
-    | David Limb                    |                       |
-    | David Bowe                    |                       |
-    | Smith Clark                   | Marketing             |
-    | Ted Kennedy                   |                       |
-    | Ronald Reagan                 |                       |
-    | Franlkin Rossevelt            |                       |
-    | Sam Malone                    |                       |
-    | Marry Georgia                 |                       |
-    | Tim Doe                       | Human Resources       |
-    | Tim Malone                    | Purchasing            |
-    |                               | Shipping              |
-    +-------------------------------+-----------------------+
+    +--------------------+-----------------------+
+    | employee           | manager_of_department |
+    +--------------------+-----------------------+
+    | John Doe           | Administration        |
+    | John Smith         | IT                    |
+    | Sean Lee           |                       |
+    | Rebecca Sean       |                       |
+    | Tim Ducan          |                       |
+    | Robert Duval       |                       |
+    | Clint Johnson      |                       |
+    | Sarrah Mcmillan    |                       |
+    | David Limb         |                       |
+    | David Bowe         |                       |
+    | Smith Clark        | Marketing             |
+    | Ted Kennedy        |                       |
+    | Ronald Reagan      |                       |
+    | Franklin Rossevelt |                       |
+    | Sam Malone         |                       |
+    | Marry Georgia      |                       |
+    | Tim Doe            | Human Resources       |
+    | Tim Malone         | Purchasing            |
+    |                    | Shipping              |
+    +--------------------+-----------------------+
     SELECT 19 rows in set (... sec)
 
 Join Conditions
